+++ fritzconnection/lib/fritzstatus.py
@@ -2,11 +2,18 @@
 Module to read status-information from an AVM FritzBox.
 """
 
-
 from __future__ import annotations
 
+import datetime
 from collections import namedtuple
-from warnings import warn
+
+from fritzconnection.core.processor import (
+    Storage,
+    InstanceAttributeFactory,
+    processor,
+    process_node,
+)
+from fritzconnection.core.utils import get_xml_root
 
 from .fritzbase import AbstractLibraryBase
 from .fritztools import (
@@ -17,9 +24,9 @@ from .fritztools import (
 )
 
 DefaultConnectionService = namedtuple(
-    "DefaultConnectionService",
-    "prefix connection_service postfix"
+    "DefaultConnectionService", "prefix connection_service postfix"
 )
+
 
 def _integer_or_original(value):
     """
@@ -30,6 +37,44 @@ def _integer_or_original(value):
         return int(value)
     except ValueError:
         return value
+
+
+@processor
+class Event:
+    """
+    Represents an AVM DeviceLog entry with the subnodes given as
+    class-attributes. The default types are strings as extracted from
+    the xml-source.
+    """
+
+    id = None
+    group = None
+    date = None
+    time = None
+    msg = None
+
+    @property
+    def datetime(self):
+        return datetime.datetime.strptime(f"{self.date}{self.time}", "%d.%m.%y%H:%M:%S")
+
+
+class DeviceLog(Storage):
+    """
+    The AVM DeviceLog is a list of Event-nodes stored in the `events`
+    instance attribute. But instances of DeviceLog are also iterables
+    for the events.
+    """
+
+    Event = InstanceAttributeFactory(Event)
+
+    def __init__(self, root):
+        self.events = list()
+        super().__init__(self.events)
+        process_node(self, root)
+
+    def __iter__(self):
+        for event in self.events:
+            yield event
 
 
 class FritzStatus(AbstractLibraryBase):
@@ -107,15 +152,6 @@ class FritzStatus(AbstractLibraryBase):
         return status["NewUptime"]
 
     @property
-    def uptime(self) -> int:
-        """
-        .. deprecated:: 1.9.0
-           Use :func:`connection_uptime` instead.
-        """
-        warn('This method is deprecated. Use "connection_uptime" instead.', DeprecationWarning)
-        return self.connection_uptime
-
-    @property
     def device_uptime(self) -> int:
         """Device uptime in seconds."""
         status = self.fc.call_action("DeviceInfo1", "GetInfo")
@@ -124,7 +160,7 @@ class FritzStatus(AbstractLibraryBase):
     @property
     def str_uptime(self) -> str:
         """Connection uptime in human-readable format."""
-        mins, secs = divmod(self.uptime, 60)
+        mins, secs = divmod(self.connection_uptime, 60)
         hours, mins = divmod(mins, 60)
         return "%02d:%02d:%02d" % (hours, mins, secs)
 
@@ -139,9 +175,7 @@ class FritzStatus(AbstractLibraryBase):
             value = status["NewX_AVM_DE_TotalBytesSent64"]
         except KeyError:
             # fallback for older FritzOS Versions not providing a 64 bit value:
-            status = self.fc.call_action(
-                "WANCommonIFC1", "GetTotalBytesSent"
-            )
+            status = self.fc.call_action("WANCommonIFC1", "GetTotalBytesSent")
             value = status["NewTotalBytesSent"]
         return _integer_or_original(value)
 
@@ -156,9 +190,7 @@ class FritzStatus(AbstractLibraryBase):
             value = status["NewX_AVM_DE_TotalBytesReceived64"]
         except KeyError:
             # fallback for older FritzOS Versions not providing a 64 bit value:
-            status = self.fc.call_action(
-                "WANCommonIFC1", "GetTotalBytesReceived"
-            )
+            status = self.fc.call_action("WANCommonIFC1", "GetTotalBytesReceived")
             value = status["NewTotalBytesReceived"]
         return _integer_or_original(value)
 
@@ -324,10 +356,7 @@ class FritzStatus(AbstractLibraryBase):
         # check for the corresponding action
         # whether mesh is supported
         try:
-            return (
-                "X_AVM-DE_GetMeshListPath"
-                in self.fc.services["Hosts1"].actions
-            )
+            return "X_AVM-DE_GetMeshListPath" in self.fc.services["Hosts1"].actions
         except KeyError:
             # can happen if "Hosts1" is not known
             return False
@@ -345,6 +374,43 @@ class FritzStatus(AbstractLibraryBase):
         """
         return ArgumentNamespace(self.fc.call_action("DeviceInfo1", "GetInfo"))
 
+    def get_avm_device_log(self, filter: str | None = None) -> DeviceLog:
+        """
+        The avm device log is a list of events with the attributes `id`,
+        `group`, `date`, `time` and `msg` holding information like "DSL
+        synchronization starting (training)" and other system messages.
+        The Method returns a DeviceLog instance holding a list of Event
+        instances. The DeviceLog instance is an iterable and can be
+        used on a FritzStatus instance like:
+
+        >>> device_log = fritzstatus.get_avm_device_log()
+        >>> for event in device_log:
+        >>>     print(event.datetime, event.msg)
+
+        The returned events can be filtered by groups like 'sys', 'net',
+        'fon', 'wlan' or 'usb'. To filter by a group provide the
+        group-name as filter-argument.
+        """
+        result = self.fc.call_action("DeviceInfo1", "X_AVM-DE_GetDeviceLogPath")
+        path = result["NewDeviceLogPath"]
+        if filter:
+            path = f"{path}&filter={filter}"
+        url = f"{self.fc.address}:{self.fc.port}{path}"
+        root_node = get_xml_root(url, session=self.fc.session)
+        return DeviceLog(root_node)
+
+    def get_cpu_temperatures(self) -> list[int]:
+        """
+        Returns a list of the last measured cpu-temperatures (Celsius).
+        The most recent entry is the first one in the list.
+        NOTE: this function call is experimental as it is based on a
+        non-public API.
+        """
+        url = f"{self.fc.http_interface.router_url}/query.lua"
+        payload = {"CPUTEMP": "cpu:status/StatTemperature"}
+        response = self.fc.http_interface.call_url(url, payload)
+        return list(map(int, response.json()["CPUTEMP"].split(",")))
+
     def get_default_connection_service(self) -> DefaultConnectionService:
         """
         Returns a namedtuple of type DefaultConnectionService:
@@ -352,14 +418,11 @@ class FritzStatus(AbstractLibraryBase):
         `device_connection` -> str (like "WANPPPConnection")
         `postfix` -> str
         """
-        result = self.fc.call_action(
-                "Layer3Forwarding1", "GetDefaultConnectionService"
-        )
-        prefix, connection_service, postfix = \
-            result["NewDefaultConnectionService"].split('.', 2)
-        return DefaultConnectionService(
-            prefix, connection_service, postfix
-        )
+        result = self.fc.call_action("Layer3Forwarding1", "GetDefaultConnectionService")
+        prefix, connection_service, postfix = result[
+            "NewDefaultConnectionService"
+        ].split(".", 2)
+        return DefaultConnectionService(prefix, connection_service, postfix)
 
     @property
     def connection_service(self) -> str:
@@ -388,7 +451,8 @@ class FritzStatus(AbstractLibraryBase):
     @property
     def has_wan_support(self) -> bool:
         """
-        True is the device supports a WAN interface.
+        True if the device supports a WAN interface.
         False otherwise.
         """
         return "Layer3Forwarding1" in self.fc.services
+