+++ docs/sources/version_history.rst
@@ -4,6 +4,32 @@ Version History
 ===============
 
 
+1.15.0-dev
+----------
+
+- support added for Python 3.14
+- FritzConnection:
+
+  - additional argument `redact_debug_log` (default: False) for optional redacting response in debug output. (#238) (#241)
+  - new method `get_cpu_temperatures()` providing a list of the last recent cpu-temperatures. (Rewrite of #232).
+
+- FritzCall:
+
+  - new attribute `Path` for the class `Call` to access an optional phone message. (#231)
+
+- FritzStatus:
+
+  - new method `get_avm_device_log()` to access system events. Requires FritzOS 8. (#234)
+
+- bugfix: in rare cases a session id of None has prevented a successfull request of the http-interface.
+
+- **removed**:
+
+  - `fritzconnection.lib.fritzhomeauto.FritzHomeAutomation.device_informations()`, deprecated in 1.9.0
+  - `fritzconnection.lib.fritzstatus.FritzStatus.uptime()`, deprecated in 1.9.0
+  - `fritzconnection.lib.fritzwlan.FritzWLAN.channel_infos()`, deprecated in 1.9.0
+
+
 1.14.0 - 2024-08-12
 -------------------
 
@@ -23,7 +49,7 @@ Version History
 - bugfix: some devices may not return system-information the propper way, causing errors on the cli output. In these cases the system-information will get ignored. (#214)
 - documentation: some typos corrected. (#202, #204)
 - testing: `tox.ini` removed because of change to `nox`. Change from `pylint` to `ruff` for linting.
-- deprecation: use of the `json` cache-format is discouraged. Use the default pickle-format instead. The highly dynamic TR-064 parser may get an ouverhaul in the future and to reduce the complexity of the parser the support of `json` for caching will be removed.
+- **deprecation**: use of the `json` cache-format is discouraged. Use the default pickle-format instead. The highly dynamic TR-064 parser may get an ouverhaul in the future and to reduce the complexity of the parser the support of `json` for caching will be removed.
 
 
 1.13.2 - 2023-09-17
@@ -96,7 +122,7 @@ Version History
 
   - bugfix: create new socket on lost connection. (#179)
 
-- Deprecations:
+- **Deprecations**:
 
   - `fritzconnection.lib.fritzhomeauto.FritzHomeAutomation.device_information()`
 
@@ -186,7 +212,8 @@ Version History
   - New method `channel_info()` (#131)
 
 - FritzHomeAutomation: New method `device_information()` (#131)
-- Deprecations:
+
+- **Deprecations**:
 
   - `fritzconnection.lib.fritzhomeauto.FritzHomeAutomation.device_informations()`
   - `fritzconnection.lib.fritzstatus.FritzStatus.uptime()`